+++ packages/hono-takibi/src/generators/hono/openapi/component/allof/process/process-alllof.ts
@@ -1,7 +1,7 @@
-import type { Config } from '../../../../../config'
-import type { Schema } from '../../../../../types'
-import { isNullableSchema } from '../../../../../core/validator/is-nullable-schema'
-import { generateZodSchemaFromSubSchema } from '../../../../zod/sub/generate-zod-schema-from-sub-schema'
+import type { Config } from '../../../../../../config'
+import type { Schema } from '../../../../../../types'
+import { isNullableSchema } from '../../../../../../core/validator/is-nullable-schema'
+import { generateZodSchemaFromSubSchema } from '../../../../../zod/sub/generate-zod-schema-from-sub-schema'
 
 type Accumulator = {
   nullable: boolean
