+++ packages/hono-takibi/src/generators/openapi/components/allof/generate-allof-code.ts
@@ -1,6 +1,6 @@
 import type { Schema } from '../../../../types'
 import type { Config } from '../../../../config'
-import { processAllOf } from './process/process-alllof'
+import { processAllOf } from '../../../hono/openapi/component/allof/process/process-alllof'
 import { generateZodIntersection } from '../../../zod/generate-zod-intersection'
 
 /**
